Update Basic Metadata and Core Metadata sections of "Defining Metadata in the Model" page to reflect Hyrax 1.0-stable accurately; closes #126.
In Hyrax 1, "Core" Metadata is possibly more accurately referred to as "Required" Metadata since that is the name of Hyrax 1 class that contains it.
However, I left the "core" language as it was since I thought "required metadata" might be ambiguous and since that is how Hyrax 2 will refer to it.