Amending documentation - a proposal and clarity
In reviewing the documentation, I'm checking to see the to be removed
documentation was written based on experience or based on other
documentation (See https://github.com/samvera/hyrax/pull/2150 for
reference and possible other conversations).

If the documentation reflects experience, then I believe that is a bug;
I would expect the `.activate!` method to be idempotent.