Adding a placeholder Rake task for Travis
We are aspiring to check the built website as part of the pull request
process. However, there are two steps:

1) Activating Travis
2) Having Travis check the Website

As of this commit we have activated travis, but do not have a task for
checking the website (that is part of PR #154). This is a placeholder
solution until we resolve the technical issues of the other PR.

Related to #154