Guide for storing fedora binaries on filesystem
The long version of this guide is in a google doc.
We do not have the bandwidth at this time to re-write
the entire guide, but I am attempting here to provide
enough guidance that an institutions wishing to undertake
a similar project can understand the issues and general
scope of the work, along with some examples of previous
art.