Adds a hook to copy generated pages into the project
This is to get around Github page not honoring the 'keep_page' directive in Jekyll config.

Normally, you could specify dynamically generated pages to the destination directory in 'keep_pages'
and those files won't get clobbered when the site is re-build, but Github Pages seems to clobber
them anyway.

So here we add a bit of logic to copy generated files form the destination directory to the source
directory, as html, so that Github will copy them over to the destination directory as-is.